.github: Fix Pull Request template links
The link to the conventions document is correct when reading it from the
template files, but does not render properly in the browser, because we
need to tell GitHub we're looking at a blob from the master branch.

The main purpose for this template is to be rendered in the browser, so
this is where the link should work correctly. Fix the URL to point to
the right document, and add a second link to the English version of the
document.

Signed-off-by: Quentin Monnet <qmo@qmon.net>